Feature/RDM-19 przebudowa seedow pod produkcje

* Added new fields to user factory

* Created user admin seeder

* Created production seeders

+++ database/seeders/Other/DocumentCategorySeeder.php
@@ -4,6 +4,7 @@ namespace Database\Seeders\Other;
 
 use App\Services\Documents\Models\DocumentCategory;
 use Illuminate\Database\Seeder;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Schema;
 
 class DocumentCategorySeeder extends Seeder
@@ -15,14 +16,22 @@ class DocumentCategorySeeder extends Seeder
      */
     public function run()
     {
-        Schema::disableForeignKeyConstraints();
-        DocumentCategory::truncate();
-        Schema::enableForeignKeyConstraints();
+        if(!App::environment('production')){
+            Schema::disableForeignKeyConstraints();
+            DocumentCategory::truncate();
+            Schema::enableForeignKeyConstraints();
+        }
 
         DocumentCategory::reguard();
 
         foreach($this->documentCategories() as $documentCategory){
-            DocumentCategory::create($documentCategory);
+
+            if(!App::environment('production')){
+                $documentCategoryModel = DocumentCategory::where('name->en', $documentCategory['name']['en'])->first();
+            }
+            if(!empty($documentCategoryModel)){
+                DocumentCategory::create($documentCategory);
+            }
         }
 
     }
@@ -30,10 +39,22 @@ class DocumentCategorySeeder extends Seeder
     private function documentCategories(): array
     {
         return [
-            ['name' => 'Offers'],
-            ['name' => 'Financial'],
-            ['name' => 'Builders'],
-            ['name' => 'Orders'],
+            ['name' => [
+                'en' => 'Offers',
+                'de' => 'Angebote'
+            ]],
+            ['name' => [
+                'en' => 'Financial',
+                'de' => 'Finanziell'
+            ]],
+            ['name' => [
+                'en' => 'Builders',
+                'de' => 'Bauherrinnen'
+            ]],
+            ['name' => [
+                'en' => 'Orders',
+                'de' => 'Aufträge'
+            ]],
         ];
     }
